fix conditional binding errors
This code built in Xcode prior to 14.3 (I don't know which version
specifically). But in 14.3, I kept getting the error

`Initializer for conditional binding must have Optional type, not 'String
.SubSequence' (aka 'Substring')` on lines 91-92.
It happens that once userIdSplit is calculated it doesn't have any
optionals so they don't need to be bound outside of the block. So I
moved the username and homeserver inside the block being processed to be
 used later.

 For a block this short, one might get away with simply using the
 userIdSplit values directly but having them named makes the code more
 understandable.